% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_novelty.R
\name{fit_novelty}
\alias{fit_novelty}
\title{Fit Model for detecting novelty effect from Averate Treatment Effect}
\usage{
fit_novelty(ate, alpha = 0.35, gamma = 2)
}
\arguments{
\item{ate}{Averate Treatment Effect, i.e. Expecttaion of Treatment - Expecttaion of Control}

\item{alpha}{parameter alpha}

\item{gamma}{parameter gamma}
}
\value{
linear model object
}
\description{
Fit Model for detecting novelty effect from Averate Treatment Effect
}
\examples{
\dontrun{
ate <- rev(1.3^(rnorm(14, 0, 0.3) + 1:14) / 10)
fit <- fit_novelty(ate)
summary(fit)
fit <- fit_novelty(ate, alpha = 0.3, gamma = 2)
summary(fit)
}

}
\seealso{
\href{https://dl.acm.org/citation.cfm?id=3291000}{How A/B Tests Could Go Wrong: Automatic Diagnosis of Invalid Online Experiments, WSDM2019}
}
